chore: move signal types to separate file
