+++ packages/qwik/src/core/shared/jsx/jsx-runtime.ts
@@ -8,7 +8,8 @@ import { logOnceWarn, logWarn } from '../utils/log';
 import { ELEMENT_ID, OnRenderProp, QScopedStyle, QSlot, QSlotS } from '../utils/markers';
 import { qDev, seal } from '../utils/qdev';
 import { isArray, isObject, isString } from '../utils/types';
-import { WrappedSignal, WrappedSignalFlags } from '../../signal/signal';
+import { WrappedSignal } from '../../signal/signal';
+import { WrappedSignalFlags } from '../../signal/types';
 import type { DevJSX, FunctionComponent, JSXNode, JSXNodeInternal } from './types/jsx-node';
 import type { QwikJSX } from './types/jsx-qwik';
 import type { JSXChildren } from './types/jsx-qwik-attributes';
