+++ packages/qwik/src/core/shared/shared-container.ts
@@ -1,7 +1,7 @@
 import type { ContextId } from '../use/use-context';
 import { trackSignalAndAssignHost } from '../use/use-core';
 import { version } from '../version';
-import type { SubscriptionData } from '../signal/signal';
+import type { SubscriptionData } from '../signal/subscription-data';
 import type { Signal } from '../signal/signal.public';
 import type { ISsrNode, StreamWriter, SymbolToChunkResolver } from '../ssr/ssr-types';
 import type { Scheduler } from './scheduler';
