+++ packages/qwik/src/core/signal/subscriber.ts
@@ -1,14 +1,10 @@
-import { QBackRefs } from '../shared/utils/markers';
-import {
-  type Consumer,
-  EffectProperty,
-  type EffectSubscription,
-  EffectSubscriptionProp,
-} from './signal';
-import type { ISsrNode } from '../ssr/ssr-types';
-import { _EFFECT_BACK_REF } from '../signal/flags';
 import { isServer } from '@qwik.dev/core/build';
+import { QBackRefs } from '../shared/utils/markers';
+import { _EFFECT_BACK_REF } from '../signal/flags';
+import type { ISsrNode } from '../ssr/ssr-types';
 import { BackRef } from './signal-cleanup';
+import type { Consumer, EffectProperty, EffectSubscription } from './types';
+import { EffectSubscriptionProp } from './types';
 
 export function getSubscriber(
   effect: Consumer,
