+++ packages/qwik/src/core/signal/subscription-data.ts
@@ -1,0 +1,19 @@
+import type { Signal } from './signal.public';
+
+export interface NodePropData {
+  $scopedStyleIdPrefix$: string | null;
+  $isConst$: boolean;
+}
+
+export interface NodePropPayload extends NodePropData {
+  $value$: Signal<unknown>;
+}
+
+/** @internal */
+export class SubscriptionData {
+  data: NodePropData;
+
+  constructor(data: NodePropData) {
+    this.data = data;
+  }
+}
