+++ packages/qwik/src/core/tests/container.spec.tsx
@@ -19,7 +19,7 @@ import { createComputed$, createSignal } from '../signal/signal.public';
 import { constPropsToSsrAttrs, varPropsToSsrAttrs } from '../ssr/ssr-render-jsx';
 import { type SSRContainer } from '../ssr/ssr-types';
 import { _qrlSync } from '../shared/qrl/qrl.public';
-import { SignalFlags } from '../signal/signal';
+import { SignalFlags } from '../signal/types';
 
 describe('serializer v2', () => {
   describe('rendering', () => {
