+++ packages/qwik/src/core/use/use-resource.ts
@@ -11,7 +11,7 @@ import { delay, isPromise, safeCall } from '../shared/utils/promises';
 import { isFunction, isObject } from '../shared/utils/types';
 import { StoreFlags, createStore, getStoreTarget, unwrapStore } from '../signal/store';
 import { useSequentialScope } from './use-sequential-scope';
-import { EffectProperty, isSignal } from '../signal/signal';
+import { isSignal } from '../signal/signal';
 import type { Signal } from '../signal/signal.public';
 import { clearAllEffects } from '../signal/signal-cleanup';
 import { ResourceEvent } from '../shared/utils/markers';
@@ -19,6 +19,7 @@ import { assertDefined } from '../shared/error/assert';
 import { noSerialize } from '../shared/utils/serialize-utils';
 import { ChoreType } from '../shared/util-chore-type';
 import { getSubscriber } from '../signal/subscriber';
+import { EffectProperty } from '../signal/types';
 
 const DEBUG: boolean = false;
